+++ projetofullstack-front/src/index.css
@@ -5,6 +5,9 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-image: url("./img/bground_5.jpg");
+  background-size: 100vw;
+  background-repeat: no-repeat;
 }
 
 code {
